Create ui and handler for adding new group
Fixes #1

Did so by adding an option at the top of select to add a new group. Changed the
way options were rendered so other ones are cleared and appended to this.

Unit tests for the symbol add group event when new button clicked. Added a unit
test to ensure no duplicate group <option> elements.

Other:
- Fixed all attr for selectors so they follow standards
- Fixed references to attr selectors so the use string format.

TODO
- Implement code that will show group add form or some sort of ui to actually
add a group.