Add a new group as an option and select it
Fixes #10

When a group is added the add UI will add it as an option and select it. The
new select group function will make no attempt to see if the group its trying to
select is valid.

There could be a race condition where this ui will not add and select the group
before the add group form hides itself. This will likely not be a problem as
there will be a css transition to hide the add group form.