Button and trigger to delete investment from list
Fixes #18

Button click and trigger to delete a list entry.

Problem discovered with Flight: on handlers in after init are not registered on
non-present elements, such as the ones here because they are not there yet but
will be rendered in templates. Got around it by using jQuery's on method which
uses delegate. This required more fixes as well. Should find better way.