Renamed field names and structure to match yql
The field names from yql were not matching what I had originally,
so I renamed the fields that exist in YQL. I also changed some fiels
out.

The structure was also different because some yql fields are an
object with a "content" attribute and others are just the value. I'm
using a function to transform them to all have an object with a
content attribute.